Fix macos zip generation following #353.
The `execute` method was not called any more on the nested task in
`AddResourcesAndZipMacApp` so the nested task was not being executed
any more.  Digging deeper, it seems like `execute` does not exist
more.

The workaround is to split the task into tasks, one depending on each
other.  This fixes #359.