Fix #350
It is possible that the database may contain a reference to a match
that has no entry in `MATCHDETAILS`, in which case an
`ArrayIndexOutOfBoundsException` is thrown when trying to display a
player with a ref to that match.

This fix tries to re-download the match details if it realises they
are not present in the db.  If ultimately, no details can be found,
the match is simply skip.

As a result, there may also be an NPE, as the date of the match is
also not present, so this fix also checks whether the match date is
present when we try to calculate the age of the player at the time.