Simplify league table and re-arrange league details screen
The league table has now been simplified: all the details about home and
away have been removed, as well as the promotion / demotion zones in the
table.  Form streak is now displayed as circles.

The league history has also been cleaned up and moved beside the league
table.  It is now bigger, and the font size honours user preferences
rather adjusting to the size of the graph.  Finally, the current team's
line now has a thicker stroke, rather than two lines drawn to make it
look thicker.

All the new colours are registered with the ThemeManager in prevision of
new future themes.