Remove check to see if change is still adjusting. (#495)
Waiting until the change is fully adjusted causes a small delay in
sync'ing the name table with the details table on the Squad table,
which is even more pronounced in dark mode, where painting is a bit
slower.  This commit removes the wait to make the change immediately.
This closes #465.

Also, the ivar `spielerUebersichTable` has been renamed
`playerOverviewTable` to reduce confusion.