Ensure the HOMainFrame does not call itself whilst still loading. (#598)
Under certain conditions HOMainFrame constructor ends up calling
`HOMainFrame.instance()` which in turn calls the HOMainFrame
constructor, as the null check of the singleton pattern is still
null (the constructor has not completed yet).

This PR ensures the main frame is not loading before display the OAuth
dialog is created.  If it still is loading, the dialog is created with
a null window, which should only cause the dialog to not be centered.