Remove player state colours. (#619)
As the state colours have now been removed, the multiline player label
now displays details about their state: injured, suspended, two yellow
cards, etc.  The single line labels do not display those icons.

This commit was originally e96d911f8edeabfb6b4cf9b18db1541281ae769b,
but it was reverted as it was causing an NPE because of a change for
the groups icons.  This commit has been cherry-picked, with the
exception of the change in question.