Calculate width with coords of left and top of current screen. (#630)
When calculating the width to be saved in the user parameters upon
exiting HO, the width of the screen is calculated as the min of the
current width and the difference between the width of the current
screen minus the current X location of the screen.

In a setup where multiple screens are present, the leftmost X may not
be 0 (e.g. X = 1096), causing the difference `width_screen -
current_pos_x` to be ridiculously small, and being stored as the
current width. (Ditto for height)

This commit offsets the difference with the leftmost X of the screen
to take into account multiple screens.

This commit also:
- translates some the comments in German into English;
- removes the deprecated user parameter `bestPostWidth` which seems to
never be read.