Automated release pipeline.
For the pipeline to kick off, a tag must be created:
- v<major>.<minor>.<patch>-dev for a dev release;
- v<major>.<minor>.<patch>-beta for a beta release;
- v<major>.<minor>.<patch> for a stable release.

It creates an automated release in Github based on a the version of
the Gradle project.  All the artefacts under `build/artefacts/` are
attached to the build.

The build process requires two secrets to be installed in the
repository:

- A github token `RELEASE_TOKEN`, with release write permissions;
- A valid install4j license `INSTALL4J_LICENSE`.