Do not return `true` when the application first starts (#1758)
Fixes #1757, although this should probably be further improved by ensuring that
if no column is selected, at least a default set of “core” columns still appear,
like nationality, etc.

This was caused by the fact that there was no `UserColumn` to be displayed, as
by default all columns are marked as `canDisplay` `false`.  This fix ensures
that upon first start all columns are displayed, therefore the
`UserColumnsTable` gets populated.  Subsequently the user can select the columns
to show or hide.  Again we should either prevent the user from de-selecting
everything, or not blow up when the user selects nothing (probably the better
solution).