Change order of Gradle tasks.
The `clean` task is called after `compileKotlin`, and before `compileJava`,
causing the output of the Kotlin compilation to be wiped, and therefore not made
available to the Java compilation.

This commit re-orders tasks to get `clean` done before the other tasks for
`compileJava`.

This commit also verifies the presence of the `INSTALL4J_HOME` env var to avoid
compile issue when checking out the project the first time.