Refactor prepared statement caching logic.
This commit introducts a `StatementCache` that manages the cached
`PreparedStatement`s in a straightforward and consistent manner; the statements in this cache can
easily be cleared out, and the cache can also be disabled.  It also keeps tracks of statistics
about the prepared statements that can be useful for debugging and tracking
commonly used ones.

This also removes some of the ad-hoc attempts at achieving similar
caching (e.g. in `TrainingsTable`, `DBUpdater`, etc.).

Finally, it also deletes the `XMLExporter` class, which has now become obsolete,
per discussion: https://github.com/ho-dev/HattrickOrganizer/discussions/1942