#1886 improve training planning (#1970)
* #1962 refactoring PlayerSkill class->enum

* #1962 refactoring PlayerSkill class->enum

* #1962 refactoring PlayerSkill: remove PlayerSkillColumn.getSkill

* #1962 refactoring PlayerSkill: remove HTSkillID

* #1962 refactoring PlayerSkill: remove redundant Skills.getSkillValue

* #1962 refactoring PlayerSkill.toString -> getLanguageString

* #1962 refactoring remove PastSkillChange.java

* hsqldb 2.7.2

* Refactor prepared statement caching logic.

This commit introducts a `StatementCache` that manages the cached
`PreparedStatement`s in a straightforward and consistent manner; the statements in this cache can
easily be cleared out, and the cache can also be disabled.  It also keeps tracks of statistics
about the prepared statements that can be useful for debugging and tracking
commonly used ones.

This also removes some of the ad-hoc attempts at achieving similar
caching (e.g. in `TrainingsTable`, `DBUpdater`, etc.).

Finally, it also deletes the `XMLExporter` class, which has now become obsolete,
per discussion: https://github.com/ho-dev/HattrickOrganizer/discussions/1942

* #1886 (wip)

* #1886 merge

* #1886 PlayerDetailPanel loads correct skill training priority

* #1886 fix FuturePlayerTraining.cut (create unit test)

* #1886 fix FuturePlayerTraining.cut (create unit test)

* #1886 delete TrainingPlanPanel.java

* #1886 fix FuturePlayerTrainingTable.storeFuturePlayerTrainings

---------

Co-authored-by: Sébastien Le Callonnec <sebastien@weblogism.com>