Check if un-managed HO install when downloading updates. (#2010)
* Do not store `updatesUrl` when no media ID (dev mode)

* Check if un-managed HO install when downloading updates.

HO now downloads an update in the browser rather than using install4j update
mechanism if the install is un-managed (i.e. has been done manually, or is a
development environment).

This also fixes an issue with the version number not set correctly in the
install4j generated artefacts, which causes issues with updates.