Main Frame code cleanup.
The key change in this class is to make actions lambdas directly in the `addActionListener`,
removing the `ActionListener` responsibility from the frame itself.  The menu items no longer need
to be ivars, and were made local variables in the `initMenu` method.

One future improvement will be to move the business logic (`selectedPlayer`) away from this
component, to clearly separate model and view.